refactor(ui): extract shared types, hooks, and helpers from App.tsx
- Create lib/types.ts with canonical TodoItem, TokenUsage, ConfiguredModelInfo,
  ResumeSessionOption, QueuedMessage, SessionUiCacheEntry types
- Extract usePanelLayout hook (~310 lines of terminal/file explorer
  position, resize, drag-to-reposition, combined panel, terminal tabs)
- Extract useMobileSidebar hook (~90 lines of swipe gesture + overflow)
- Extract message-helpers.ts (~190 lines of toRelayMessage, deduplicateMessages,
  normalizeMessages, normalizeModel, normalizeSessionName, augmentThinkingDurations,
  normalizeModelList)
- Extract ShortcutsDialog component

App.tsx shrinks from 4,368 to 3,734 lines (-634 lines, -15%).
Removes 3x duplicated TodoItem, 2x TokenUsage, 2x ResumeSessionOption.
All pre-existing tests pass, typecheck clean.
